+++ src/main/java/com/example/newsfeed/dto/menudto/MenuRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.newsfeed.dto;
+package com.example.newsfeed.dto.menudto;
 
 import lombok.Getter;
