+++ src/main/java/com/example/newsfeed/dto/menudto/MenuResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.newsfeed.dto;
+package com.example.newsfeed.dto.menudto;
 
 import com.example.newsfeed.entity.Comment;
 import com.example.newsfeed.entity.Menu;
