+++ src/main/java/com/example/newsfeed/dto/userdto/PwdCheckRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.newsfeed.dto;
+package com.example.newsfeed.dto.userdto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
