+++ src/main/java/com/example/newsfeed/dto/userdto/UserRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.newsfeed.dto;
+package com.example.newsfeed.dto.userdto;
 
 import lombok.Getter;
